+++ dcti-api/src/test/java/com/dwarfeng/dcti/api/integration/configuration/ServiceExceptionMapperConfiguration.java
@@ -1,0 +1,20 @@
+package com.dwarfeng.dcti.api.integration.configuration;
+
+import com.dwarfeng.subgrade.impl.exception.MapServiceExceptionMapper;
+import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
+import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.Map;
+
+@Configuration
+public class ServiceExceptionMapperConfiguration {
+
+    @Bean
+    public MapServiceExceptionMapper mapServiceExceptionMapper() {
+        Map<Class<? extends Exception>, ServiceException.Code> des = ServiceExceptionHelper.putDefaultDestination(null);
+        des = com.dwarfeng.dcti.sdk.util.ServiceExceptionHelper.putDefaultDestination(des);
+        return new MapServiceExceptionMapper(des, com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.UNDEFINED);
+    }
+}